postgres/pyproject.toml
[project]
name = "postgres"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.30.0",
    "python-dotenv>=1.0.1",
]

[tool.uv.sources]
